\name{cosmic.cancer.gene.census-package}
\alias{cosmic.cancer.gene.census-package}
\alias{cosmic.cancer.gene.census}
\docType{package}
\title{
COSMIC cancer gene census R package.
}
\description{
Contains a dataframe of HGNC symbols of genes found in the COSMIC Cancer Gene Census.
}
\details{
\tabular{ll}{
Package: \tab cosmic.cancer.gene.census\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2016-07-13\cr
License: \tab GPL-2\cr
}
filter.cosmic.cancer.gene.census()

import.comis.cancer.gene.census.data()

cosmic.cancer.gene.census
}
\author{
Richard de Borja <richard.deborja@sickkids.ca>

Maintainer: Richard de Borja <richard.deborja@sickkids.ca>
}
\references{
http://cancer.sanger.ac.uk/cancergenome/projects/census/
}
\keyword{ COSMIC }
\keyword{ cancer }
\keyword{ census }
